+++ __tests__/OmnichannelChatSDK.web.spec.ts
@@ -38,6 +38,7 @@ describe('Omnichannel Chat SDK (Web)', () => {
         const chatSDK = new OmnichannelChatSDK(omnichannelConfig);
         chatSDK.getChatConfig = jest.fn();
         
+        chatSDK.liveChatVersion = 1;
         await chatSDK.initialize();
 
         chatSDK.IC3Client = {
@@ -74,6 +75,7 @@ describe('Omnichannel Chat SDK (Web)', () => {
         const chatSDK = new OmnichannelChatSDK(omnichannelConfig);
         chatSDK.getChatConfig = jest.fn();
 
+        chatSDK.liveChatVersion = 1;
         await chatSDK.initialize();
 
         chatSDK.IC3Client = {
@@ -109,6 +111,7 @@ describe('Omnichannel Chat SDK (Web)', () => {
         chatSDK.getChatConfig = jest.fn();
         chatSDK.getChatToken = jest.fn();
         
+        chatSDK.liveChatVersion = 1;
         await chatSDK.initialize();
 
         chatSDK.OCClient.sessionInit = jest.fn();
@@ -134,6 +137,7 @@ describe('Omnichannel Chat SDK (Web)', () => {
         chatSDK.getChatConfig = jest.fn();
         chatSDK.getChatToken = jest.fn();
         
+        chatSDK.liveChatVersion = 1;
         await chatSDK.initialize();
 
         chatSDK.OCClient.sessionInit = jest.fn();
@@ -162,6 +166,7 @@ describe('Omnichannel Chat SDK (Web)', () => {
         chatSDK.getChatConfig = jest.fn();
         chatSDK.getChatToken = jest.fn();
         
+        chatSDK.liveChatVersion = 1;
         await chatSDK.initialize();
 
         chatSDK.callingOption = CallingOptionsOptionSetNumber.NoCalling;