+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Added exception on initialization failure
 - Upgraded ACSAdapter to version beta.17
 - Changed default LiveChatVersion to V2
+- Added `botSurveyInviteLink` and `botFormsProLocale` the `getPostChatSurveyContext()` response
 
 ### Fixed
 - Fix `ChatAdapterOptionalParams.ACSAdapter.options.egressMiddleware` being used as `ingressMiddleware`